Add names (usually filenames) to Input.
In a multi-file project, it's handy for a `ParserPosition` to know what
file the error was located in, not just the line and column. This commit
adds a string `user_name` to the `Input` class, which is then used by
`ParserPosition` objects. This name can be any user-meaningful name, but
it probably ought to be a filename whenever possible.