Don't add a std::string to ParserPosition.
Instead of copying strings around in a struct that gets instantiated a lot,
use the existing `Input::iterator` to get back to the `Input`, which can
give us the filename through its `name` method.

This technique requires the addition of an `Input::iterator::input()`
method. I've made this return a reference and created a `NoInput` class
for these situations, but if preferred we could instead return a nullptr
if that's what the `Input` is set to in the `Input::iterator`.